Most of updater package has been moved back to the main application. Should modify gradle task to embed the new update migrator jar and as well as change the updater back into a component manager.
