+++ src/main/java/address/controller/MainController.java
@@ -10,8 +10,7 @@ import address.model.datatypes.person.ReadOnlyViewablePerson;
 import address.model.ModelManager;
 import address.model.datatypes.person.ReadOnlyPerson;
 import address.model.datatypes.tag.Tag;
-import address.updater.UpdateProgressNotifier;
-import commons.UpdateInformationNotifier;
+import address.update.UpdateProgressNotifier;
 import address.util.*;
 import address.util.collections.UnmodifiableObservableList;
 import com.google.common.eventbus.Subscribe;
