+++ src/main/java/launcher/Launcher.java
@@ -15,7 +15,7 @@ import java.util.concurrent.Executors;
  * Will also launch the update migrator before the application to apply any pending updates
  */
 public class Launcher extends Application {
-    private static final String CLASS_PATH_REGEX = File.pathSeparator + "lib" + File.separator + "*";
+    private static final String CLASS_PATH = File.pathSeparator + "lib" + File.separator + "*";
     private static final String ERROR_LAUNCH = "Failed to launch";
     private static final String ERROR_RUNNING = "Failed to run application";
     private static final String ERROR_TRY_AGAIN = "Please try again, or contact developer if it keeps failing.";
@@ -34,26 +34,13 @@ public class Launcher extends Application {
     }
 
     private void run() throws IOException {
-        runUpdateMigratorAndWait();
         runMainApplication();
         stop();
     }
 
-    private void runUpdateMigratorAndWait() throws IOException {
-        try {
-            String command = String.format("java -ea -cp %s updater.UpdateMigrator", CLASS_PATH_REGEX);
-            System.out.println("Starting updater migrator: " + command);
-            Process process = Runtime.getRuntime().exec(command, null, new File(System.getProperty("user.dir")));
-            System.out.println("Update migrator launched");
-            process.waitFor();
-        } catch (IOException | InterruptedException e) {
-            throw new IOException(ERROR_RUNNING, e);
-        }
-    }
-
     private void runMainApplication() throws IOException {
         try {
-            String command = String.format("java -ea -cp %s address.MainApp", CLASS_PATH_REGEX);
+            String command = String.format("java -ea -cp %s address.MainApp", CLASS_PATH);
             System.out.println("Starting main application: " + command);
             Runtime.getRuntime().exec(command, null, new File(System.getProperty("user.dir")));
             System.out.println("Main application launched");
