+++ src/main/java/updater/BackupHandler.java
@@ -1,5 +1,6 @@
 package updater;
 
+import address.update.DependencyHistoryHandler;
 import commons.*;
 
 import java.io.File;
@@ -38,7 +39,7 @@ public class BackupHandler {
      * Creates a backup unless app is already being run from backup JAR.
      */
     protected void createAppBackup() throws IOException {
-        //TODO: generate this dynamically, temp for updater refactoring
+        //TODO: generate this dynamically, temp for update refactoring
         File mainAppJar = new File("lib/resource-" + currentVersion.toString() + ".jar");
 
         if (isRunFromBackupJar(mainAppJar)) {
